convert the register page to use form fields and unobtrusive validation
