+++ package.json
@@ -45,6 +45,8 @@
     "jquery": "^3.1.0",
     "jquery-migrate": "3.0.0",
     "jquery-nearest": "^1.4.0",
+    "jquery-validation": "^1.17.0",
+    "jquery-validation-unobtrusive": "^3.2.6",
     "jsonwebtoken": "^7.4.3",
     "karma-webpack": "^2.0.4",
     "moment": "^2.18.1",
